% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-sf.R
\name{cast_font_sf}
\alias{cast_font_sf}
\title{Convert font into Spatial SF object}
\usage{
cast_font_sf(font_df)
}
\arguments{
\item{font_df}{font created by \code{\link[=glyph]{glyph()}}}
}
\value{
font polygon as an SF object
}
\description{
Convert font into Spatial SF object
}
\examples{
letter_s <- glyph("s")
letter_s_sf <- cast_font_sf(letter_s)
}
