% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-sf.R
\name{cast_packing_poly}
\alias{cast_packing_poly}
\title{Cast a polygon created with \code{create_poly_packings} as an SF object}
\usage{
cast_packing_poly(poly_packing)
}
\arguments{
\item{poly_packing}{polygon created with \code{create_poly_packings} as an SF object}
}
\value{
SF object
}
\description{
Cast a polygon created with \code{create_poly_packings} as an SF object
}
\examples{
packed_poly <- create_poly_packings(areas = c(1:2), n_points = 3)
cast_packing_poly(packed_poly)
}
