% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle-height.R
\name{circle_height}
\alias{circle_height}
\alias{circle_height.sf}
\alias{circle_height.data.frame}
\title{Calculate circle height}
\usage{
circle_height(x)

\method{circle_height}{sf}(x)

\method{circle_height}{data.frame}(x)
}
\arguments{
\item{x}{circle created by \code{\link[=create_circle_plate]{create_circle_plate()}}}
}
\value{
numeric height of circle
}
\description{
Calculate circle height
}
