% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-circle-plate.R
\name{create_circle_plate}
\alias{create_circle_plate}
\title{Create a plate of circles on which the ishihara plate will be created}
\usage{
create_circle_plate()
}
\value{
data.frame of circles
}
\description{
Create a plate of circles on which the ishihara plate will be created
}
\examples{
circle_plate <- create_circle_plate()
}
