% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly-packing.R
\name{create_poly_packings}
\alias{create_poly_packings}
\title{Create polygons of a given area that do not overlap}
\usage{
create_poly_packings(areas, n_points)
}
\arguments{
\item{areas}{numeric vector of areas}

\item{n_points}{number of points to create the polygon. Three = triangle,
and about 50 or more gives a convincing circle.}
}
\value{
tibble of x, y coords of the points, and id, identifying each polygon
}
\description{
Create polygons of a given area that do not overlap
}
\examples{
# create a bunch of triangles that do not overlap
create_poly_packings(areas = c(1:2), n_points = 3)
}
