% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-helpers.R
\name{font_center}
\alias{font_center}
\title{Font Center}
\usage{
font_center(font_df)
}
\arguments{
\item{font_df}{font created by \code{\link[=glyph]{glyph()}}}
}
\value{
\code{font_df} but the x and y values are scaled on 0
}
\description{
Centers a font around 0
}
\examples{
glyph("s") \%>\% font_increase(10) \%>\% font_center()
}
