% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-helpers.R
\name{font_increase}
\alias{font_increase}
\title{Font increase}
\usage{
font_increase(font_df, increase)
}
\arguments{
\item{font_df}{font created by \code{\link[=glyph]{glyph()}}}

\item{increase}{how much to multiply font size by}
}
\value{
data.frame increased by a given increase size
}
\description{
Font increase
}
\examples{
glyph("s") \%>\% font_increase(10)
}
