% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-helpers.R
\name{font_rescale}
\alias{font_rescale}
\title{Rescale the font to fit inside the circle packings}
\usage{
font_rescale(font_df, circle_df, prop = 0.8)
}
\arguments{
\item{font_df}{font created by \code{\link[=glyph]{glyph()}}}

\item{circle_df}{circle packing data frame with \code{\link[=create_circle_plate]{create_circle_plate()}}}

\item{prop}{proportion of the circle size to cover. Default is 0.8}
}
\value{
rescaled font to cover the proportion of the height of the circle
}
\description{
Rescale the font to fit inside the circle packings
}
\examples{
letter_s <- glyph("s")
circle_plate <- create_circle_plate()
letter_s_bigger <- font_rescale(letter_s, circle_plate)
}
