% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph.R
\name{glyph}
\alias{glyph}
\title{Create a text polygon}
\usage{
glyph(glyph, family = "serif", face = "regular", nseg = 10)
}
\arguments{
\item{glyph}{text input - use only one character}

\item{family}{font family - see \code{\link[sysfonts:font_families]{sysfonts::font_families()}}}

\item{face}{Font face, possible values are "regular", "bold", "italic" and
"bolditalic".}

\item{nseg}{The number of line segments to approximate a piece of curve in
the outline. The larger nseg is, the smoother text outlines would be.
Usually a value between 5~20 would be enough.}
}
\value{
tibble with columns \code{x}, \code{y}, and \code{glyph}
}
\description{
Create a text polygon
}
\examples{
glyph("s")
}
