% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ishihara-data.R
\name{ishihara_data}
\alias{ishihara_data}
\title{Create Ishihara data}
\usage{
ishihara_data(text)
}
\arguments{
\item{text}{character vector, only one character}
}
\value{
ishihara plate with packed circles that are marked
with whether the text overlays on the circles
}
\description{
Create Ishihara data
}
\examples{
ishihara_data("s")
}
