% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay-text-in-circle.R
\name{overlay_text_in_circle}
\alias{overlay_text_in_circle}
\title{Overlay text over circle packings and identify intersections}
\usage{
overlay_text_in_circle(circle_sf, glyph_sf)
}
\arguments{
\item{circle_sf}{SF object of circle packings}

\item{glyph_sf}{SF object of text}
}
\value{
\code{circle_df} with extra column, \code{in_text} that identified whether
\code{glyph_sf} is on a circle in \code{circle_df}.
}
\description{
Overlay text over circle packings and identify intersections
}
\examples{
circle_sf <- create_circle_plate() \%>\% cast_packing_poly()
glyph_sf <- glyph("s") \%>\% cast_font_sf()
}
