\name{plate1}
\alias{plate1}
\title{Ishihara plate #1}
\description{
	Ishihara plate #1. Everyone should see the number 12 when plotted. Normally used as a demo plate.
}
\usage{
data(plate2)
}
\format{
	Object of custom object class \code{"ishihara"}.
}
\source{
	Source here.
}
\examples{
	data(plate1)
	plate1
	plot(plate1)
	print(plate1,alpha=0.5)
}
\keyword{datasets}
