\name{plate2}
\alias{plate2}
\title{Ishihara plate #2}
\description{
	Ishihara plate #2. People with normal color vision show see the number 2 when plotted.
}
\usage{
data(plate2)
}
\format{
	Object of custom object class \code{"ishihara"}.
}
\source{
	Source here.
}
\examples{
	data(plate2)
	plate2
	plot(plate2)
	print(plate2,show.hidden=TRUE)
}
\keyword{datasets}
