\name{plate9}
\alias{plate9}
\title{Ishihara plate #9}
\description{
	Ishihara plate #9. People with normal color vision show see the number \emph{74} when plotted.
}
\usage{
data(plate2)
}
\format{
	Object of custom object class \code{"ishihara"}.
}
\source{
	Source here.
}
\examples{
	data(plate9)
	plate9
	plot(plate9)
	print(plate9,alpha=0.2)
}
\keyword{datasets}
