\name{plot.ishihara}
\alias{plot.ishihara}
\alias{print.ishihara}
\title{plot & print method for "ishihara" object class}
\usage{
\method{plot}{ishihara}(x, ...)
\method{print}{ishihara}(x, ...)
}
\arguments{
	\item{x}{object of class \code{"ishihara"}.}
	\item{...}{optional arguments.}
}
\description{
	Methods to print & plot \code{"ishihara"} plates.
}
\details{
	More details of the function here.
}
\value{
	Plots or prints an \code{"ishihara"} class object.
}
\references{
	References here.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\keyword{color-vision}

