% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-helpers.R
\name{prop_circle_height}
\alias{prop_circle_height}
\title{Proportion of the height of packed circles}
\usage{
prop_circle_height(prop, circle)
}
\arguments{
\item{prop}{numeric}

\item{circle}{circle packing data frame with \code{\link[=create_circle_plate]{create_circle_plate()}}}
}
\value{
numeric of the size of the font increase
}
\description{
Proportion of the height of packed circles
}
\examples{
circle_plate <- create_circle_plate()
prop_circle_height(0.8, circle_plate)
}
